Extended /etc/issue reading and parsing
In essence, we treat /etc/issue just like agetty(8) does - with some
omissions. Specifically we are able to parse and replace a few escape
sequences like `\l`. The README has been updated to include a list of
what escape sequences are currently supported. Additionally, we have
added the option to disable the printing of /etc/issue.